requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
mysql-connector-python
google-cloud-storage
python-multipart
python-dotenv
pytest
pytest-asyncio
pytesseract
PyPDF2
torch
pillow
pdf2image
spacy